STM32F0: code alignment
No impact, no change
