+++ targets/TARGET_STM/TARGET_STM32F0/STM32Cube_FW/CMSIS/system_stm32f0xx.h
@@ -101,8 +101,6 @@ extern const uint8_t APBPrescTable[8];    /*!< APB prescalers table values */
   
 extern void SystemInit(void);
 extern void SystemCoreClockUpdate(void);
-extern void SetSysClock(void);
-
 /**
   * @}
   */
