+++ targets/TARGET_STM/TARGET_STM32F0/STM32Cube_FW/STM32F0xx_HAL_Driver/stm32f0xx_hal_adc.h
@@ -44,9 +44,6 @@
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f0xx_hal_def.h"  
 
-/* Include low level driver */
-#include "stm32f0xx_ll_adc.h"
-
 /** @addtogroup STM32F0xx_HAL_Driver
   * @{
   */
