+++ targets/TARGET_STM/TARGET_STM32F0/STM32Cube_FW/STM32F0xx_HAL_Driver/stm32f0xx_hal_def.h
@@ -45,7 +45,7 @@
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f0xx.h"
 #if defined(USE_HAL_LEGACY)
-#include "stm32_hal_legacy.h"
+  #include "Legacy/stm32_hal_legacy.h"
 #endif
 #include <stdio.h>
