+++ targets/TARGET_STM/TARGET_STM32F0/STM32Cube_FW/STM32F0xx_HAL_Driver/stm32f0xx_hal_rtc.h
@@ -43,7 +43,6 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f0xx_hal_def.h"
-#include "stm32f0xx_ll_rtc.h"
 
 /** @addtogroup STM32F0xx_HAL_Driver
   * @{
