+++ targets/TARGET_STM/TARGET_STM32F0/STM32Cube_FW/stm32f0xx_hal_conf.h
@@ -287,8 +287,7 @@
 
 /* Exported macro ------------------------------------------------------------*/
 #ifdef  USE_FULL_ASSERT
-/* ALL MBED targets use same stm32_assert.h */
-#include "stm32_assert.h"
+#include "stm32_assert.h" // MBED patch
 #else
   #define assert_param(expr) ((void)0U)
 #endif /* USE_FULL_ASSERT */
