+++ targets/TARGET_STM/TARGET_STM32F0/TARGET_STM32F070xB/system_clock.c
@@ -46,75 +46,6 @@ uint8_t SetSysClock_PLL_HSI(void);
 
 
 /**
-  * @brief  Setup the microcontroller system.
-  *         Initialize the default HSI clock source, vector table location and the PLL configuration is reset.
-  * @param  None
-  * @retval None
-  */
-void SystemInit(void)
-{
-    /* Reset the RCC clock configuration to the default reset state ------------*/
-    /* Set HSION bit */
-    RCC->CR |= (uint32_t)0x00000001U;
-
-#if defined (STM32F051x8) || defined (STM32F058x8)
-    /* Reset SW[1:0], HPRE[3:0], PPRE[2:0], ADCPRE and MCOSEL[2:0] bits */
-    RCC->CFGR &= (uint32_t)0xF8FFB80CU;
-#else
-    /* Reset SW[1:0], HPRE[3:0], PPRE[2:0], ADCPRE, MCOSEL[2:0], MCOPRE[2:0] and PLLNODIV bits */
-    RCC->CFGR &= (uint32_t)0x08FFB80CU;
-#endif /* STM32F051x8 or STM32F058x8 */
-
-    /* Reset HSEON, CSSON and PLLON bits */
-    RCC->CR &= (uint32_t)0xFEF6FFFFU;
-
-    /* Reset HSEBYP bit */
-    RCC->CR &= (uint32_t)0xFFFBFFFFU;
-
-    /* Reset PLLSRC, PLLXTPRE and PLLMUL[3:0] bits */
-    RCC->CFGR &= (uint32_t)0xFFC0FFFFU;
-
-    /* Reset PREDIV[3:0] bits */
-    RCC->CFGR2 &= (uint32_t)0xFFFFFFF0U;
-
-#if defined (STM32F072xB) || defined (STM32F078xx)
-    /* Reset USART2SW[1:0], USART1SW[1:0], I2C1SW, CECSW, USBSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFCFE2CU;
-#elif defined (STM32F071xB)
-    /* Reset USART2SW[1:0], USART1SW[1:0], I2C1SW, CECSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFFCEACU;
-#elif defined (STM32F091xC) || defined (STM32F098xx)
-    /* Reset USART3SW[1:0], USART2SW[1:0], USART1SW[1:0], I2C1SW, CECSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFF0FEACU;
-#elif defined (STM32F030x6) || defined (STM32F030x8) || defined (STM32F031x6) || defined (STM32F038xx) || defined (STM32F030xC)
-    /* Reset USART1SW[1:0], I2C1SW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFFFEECU;
-#elif defined (STM32F051x8) || defined (STM32F058xx)
-    /* Reset USART1SW[1:0], I2C1SW, CECSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFFFEACU;
-#elif defined (STM32F042x6) || defined (STM32F048xx)
-    /* Reset USART1SW[1:0], I2C1SW, CECSW, USBSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFFFE2CU;
-#elif defined (STM32F070x6) || defined (STM32F070xB)
-    /* Reset USART1SW[1:0], I2C1SW, USBSW and ADCSW bits */
-    RCC->CFGR3 &= (uint32_t)0xFFFFFE6CU;
-    /* Set default USB clock to PLLCLK, since there is no HSI48 */
-    RCC->CFGR3 |= (uint32_t)0x00000080U;
-#else
-#warning "No target selected"
-#endif
-
-    /* Reset HSI14 bit */
-    RCC->CR2 &= (uint32_t)0xFFFFFFFEU;
-
-    /* Disable all interrupts */
-    RCC->CIR = 0x00000000U;
-
-    /* Enable SYSCFGENR in APB2EN, needed for 1st call of NVIC_SetVector, to copy vectors from flash to ram */
-    RCC->APB2ENR |= RCC_APB2ENR_SYSCFGEN;
-}
-
-/**
   * @brief  Configures the System clock source, PLL Multiplier and Divider factors,
   *               AHB/APBx prescalers and Flash settings
   * @note   This function should be called only once the RCC clock configuration
@@ -154,10 +85,12 @@ void SetSysClock(void)
 /******************************************************************************/
 /*            PLL (clocked by HSE) used as System clock source                */
 /******************************************************************************/
-uint8_t SetSysClock_PLL_HSE(uint8_t bypass)
+MBED_WEAK uint8_t SetSysClock_PLL_HSE(uint8_t bypass)
 {
     RCC_OscInitTypeDef RCC_OscInitStruct = {0};
     RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
+
+    __HAL_RCC_SYSCFG_CLK_ENABLE();
 
     //Select HSI as system clock source to allow modification of the PLL configuration
     RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_SYSCLK;
@@ -208,6 +141,8 @@ uint8_t SetSysClock_PLL_HSI(void)
 {
     RCC_OscInitTypeDef RCC_OscInitStruct;
     RCC_ClkInitTypeDef RCC_ClkInitStruct;
+
+    __HAL_RCC_SYSCFG_CLK_ENABLE();
 
     // Select PLLCLK = 48 MHz ((HSI 8 MHz / 2) * 12)
     RCC_OscInitStruct.OscillatorType          = RCC_OSCILLATORTYPE_HSI;
