+++ targets/TARGET_STM/TARGET_STM32F0/TARGET_STM32F072xB/TOOLCHAIN_ARM/stm32f072xb.sct
